Rename queijo executable to lrt. (#46)
This PR renames the executable from `queijo` to `lrt` and changes the
name of the entrypoint of the CLI program to `main.cpp`. I think it's
reasonable to maintain queijo as a project-internal codename, but we can
change the folders and cmake artifact names in the future if we need to.