Add an initial collections library. (#47)
Adds an initial collections library to `std` with iterative
implementations of `map`, `filter`, and `fold`. I imagine there'll be
more useful things to add here in the future, but I figure putting down
something simple with the obvious bits is a good starting point, and
helps people get a sense of scope for lrt.