Refactor runtime libraries to be explicitly required. (#48)
This change refactors lrt's provided runtime libraries to function not
as globals, but as special requirable modules provided under the `@lrt`
alias. The libraries still provide the standard flow to open them as
globals if embedders want to use lrt. In follow-up work, we should
figure out a common interface for this functionality in luau-land and
put it all under `@std`. That way alternative runtimes can use a sort of
`std-runtime` library with a similar structure that runtime-agnostic
code can target, rather than forcing the runtime's API surface to comply
with ours if people want compatibility.