Add initial `std/time` library with a `duration` type. (#50)
This PR adds an initial version of a `std/time` library for lrt. It
currently features a `duration` data type with some useful functionality
for manipulating `durations` including via overloaded operators. Future
additions to this library should include wrapping builtin time
functionality in a type-safe interface and allowing it to interact with
`duration` properly. We should probably also deal with overflow cases
directly and produce errors (either with `error`, or we should adopt a
result type convention).