`table` and `vector` Libraries (#52)
This PR renames the original `collections` library into `table` and
re-exports all of the base Luau `table` library. Likewise, we introduce
a new `vector` library that re-exports the base Luau `vector` library
and adds some simple pure Luau extensions. This pattern of growing base
libraries feels nice to me, but it might be problematic in how it
interacts with magic functions and so we might need to use a different
pattern (perhaps with `__index`).