Clean up `@lrt/fs`. (#57)
The initial implementation of `@lrt/fs` is really hacky, and could do
with a lot of clean up both in terms of the code quality, and the actual
provided API. I'm starting by cleaning up the code itself, especially
helper functions which were written in a very C-style, and can be
rewritten to be much more natural C++ code.

There's definitely other stuff to clean up as well though, some of which
might reasonably be done in their own follow-up PRs. All of the printing
in the code should probably be replaced with luau runtime errors. I'll
probably do at least that before calling this draft ready, but other
suggestions are welcome as well.