Add a Python support script for managing cmake and ninja in one place. (#65)
This is so obviously a future opportunity for dogfooding (especially
once we get to a world where you can `brew install lute` or something),
but for now, we don't have an OS library and I hate using cmake and
ninja directly. I've thrown together `luthier` to do all the heavy
lifting. It has both cutesy names (`string`, `craft`, `play`, admittedly
the idea that you `string` before `craft` is a bit nonsense, but it's
the best I could come up with for configure) and ordinary names
(`configure`, `build`, `run`). `build` will automatically `configure` if
it hasn't been done, and `run` will automatically `configure` and
`build` as necessary.