Split some libraries from `std` into `batteries`. (#87)
In the interest of taking a more principled approach here, but still
colocating useful libraries we need until we have a package manager,
this PR splits up some of the libraries from `std` into a separate
library called `batteries`. It also updates the README to describe the
general principle behind the partition. Overall, we expect that `lute`
will contain minimally the things that must be implemented in C++ to
extend the expressivity of Luau for general-purpose programming. `std`
will contain any code that is explicitly extending or "filling out" the
rest of a library provided by C++ code (whether that's Luau builtins or
Lute's C++ runtime extensions). Finally, `batteries` will contain any
additional modules that will be useful for us in building a package
manager and additional software atop `lute` before the package manager
itself is ready.