Perform a handful of housekeeping tasks (#108)
This PR performs a bit of housekeeping on lute. First, we upgrade Luau
to the latest release. Second, we fix all the uses of `vm.create` in the
examples which were broken in the renaming effort from `vm.spawn` to
`vm.create` for consistency with the rest of luau's builtin
constructors. Finally, we redo naming in `@batteries/json` to be
consistently in camelCase for everything internal. Our standard, at
least for now, is that exposed APIs must be luacase (i.e. all
lowercase), but internal code should be camelCase. This is similar to
the standard Luau's codebase itself adopts in C++.