Rename `@batteries/cliargs` to `@batteries/cli` and clean up the code. (#109)
This PR does some basic tidying of `@batteries/cli` (after renaming it
from `@batteries/cliargs`). Clean up includes some minor changes to the
name of the interface, cleaning up some of the parsing code itself to
have easier to follow control flow, and fixing some of the incorrect
type info in the library.