Manage dependencies via luthier (#111)
We've already hit a wall with git-subrepo in #105 because subrepos
interact awfully with submodules. All of this is miserable, and I think
we're going to go the way that every mid-sized C++ project with
dependencies must go to stay sane: embracing their own bespoke tool. As
such, I've replaced the check-in for every dependency here with a
`foo.tune` file in `extern` that describes the remote source and the tag
that we would like to check out, and `luthier` now has the ability to
fetch dependencies. It will do this as part of a `configure` step, but
you can also invoke it directly with the `fetch` subcommand.