Split run and check commands (#183)
This splits lute into two subcommands. `run` and `check`. The motivation
is to support a more ergonomic `run` command while not compromising on
the existing `--check` ability to process multiple files.

```
Usage: lute <command> [options] [arguments...]

Commands:
  run (default)   Run a Luau script.
  check           Type check Luau files.

Run Options (when using 'run' or no command):
  lute [run] <script.luau> [args...]
    Executes the script, passing [args...] to it.

Check Options:
  lute check <file1.luau> [file2.luau...]
    Performs a type check on the specified files.

General Options:
  -h, --help    Display this usage message.
```

If run/check is not provided, run is inferred, and the unrecognized
command is passed to run as if it's the script.

Closes #174