Only allow `run` flags before the script (#192)
@JohnnyMorganz made a good point that if we add new flags with the
current `run` command, we can potentially break existing scripts. To fix
this, we can only accept flags before the script name on the `run`
command. This simplifies the command parsing significantly, as we don't
need to check `--` or look for `-h/--help` after the file is found.

Python, Node.js, Bun, and Deno all seem to share this behavior.