CST: Handle if statements (excluding elseif) and nils (#202)
This PR adds support for serializing a full if statement as a CST node.
For now, we exclude special handling of elseif blocks within if
statements, this will be added in a future change.

Alongside this, we serialize `nil` expressions directly as a token. This
allows us to successfully roundtrip the `writeFile.luau` example.