CST: Add initial support for anonymous functions (#203)
This PR adds support for serializing function bodies and anonymous
functions.

For now, this only includes simple function bodies with no type
annotations. We serialize the parentheses, parameters (excluding type
specifiers) and end token.

For an anonymous function, we also serialize the function token
separately. We create a custom `AstExprAnonymousFunction` node in the
Luau CST to handle this (note that in C++ land this does not exist, it
is only an `AstExprFunction` directly). This new node is added for
simplicity.

A follow up change will add serialization of type specifiers and
generics in function bodies

Co-authored-by: ariel <aweiss@hey.com>