CST: Handle repeat, break and continue (#210)
This PR adds support for serializing all the the repeat stmt node, as
well as `break` and `continue` last stmt

For break and continue, it is fairly simple - we serialize the tokens
directly.

For repeat, we serialize the `repeat` and `until` tokens using the CST
data, alongside the condition and body.