Remove restriction on the type of require results. (#227)
We originally implemented a restriction on require results being a table
or function so that scripts would intrinsically be forwards-compatible
as they evolve. To my surprise, this is a controversial choice. As part
of the as-of-yet-unwritten style guide of Lute, we'll never return
anything that is not a table or function, but users are free to do how
they see fit.