CST: Handle normal + compound assignment (#224)
This PR extends the support of the CST to handle assignment (`var =
expr`) and compound assignment (e.g. `var += expr`) statements. The main
part of this is correctly storing the operator token.

Like binary operators, we could type the operator as a union of
singleton strings representing the possible operators. For now we leave
this alone, but it can be modified in a future change.