CST: Handle parsing expressions directly (#232)
To help make writing tests easier, we want to support returning a CST
when parsing expressions directly. This PR adds CST support to the
parseexpr function (as well as adding a wrapper around it in the
batteries syntax package).

We extend the visitor set up to allow visiting an overall expression. We
also add "end" visitors. For now, this is only for expression and blocks
(as necessary for codemod tooling), however we may want to add "end"
visitors for other nodes in the future.

The printer logic is also modified to support printing expressions
directly.