Improve stack trace when debugging required modules (#238)
Required modules run in their own threads in Lute, and there isn't an
easy way to automatically "inherit" a stack trace across a thread
boundary. To make it more user-friendly to debug where an issue is
coming from in required modules, each module thread now stores its stack
trace in its error message in the case of an error, appended to the
actual error message.

---

To demonstrate, I have five Luau files that require each other in a
chain: `a.luau`, `b.luau`, `c.luau`, `d.luau`, and `e.luau`. At the end
of the chain, `e.luau` attempts to (illegally) yield.

Prior to this change, Lute outputted this:
```
module can not yield
stacktrace:
[C] function require
a:1
```

This is what Lute outputs now:
```
module can not yield
[C] function yield
./e:1
[C] function require
./d:1
[C] function require
./c:1
[C] function require
./b:1

stacktrace:
[C] function require
a:1
```

Fixes #131.