Add result to @batteries (#159)
Many operations involving the file system & networking can fail. One of
the most common ways of handling this is pcall, but this is not type
safe.

A common idiom is to have a Result type, somewhat similar to Rust's
Result type. It's an extremely common and useful idiom. Any project
written in Lute interacting with APIs which could error will end up
encountering the problem of gracefully handling failure; the result
battery significantly helps with that.

---------

Co-authored-by: ariel <aweiss@hey.com>