CST: Handle generic definitions (#268)
This PR adds the generic definition nodes into the function and type
alias definitions. This involves serializing AstGenericType /
AstGenericTypePack.

Generic definitions can be punctuated, and are separated by commas.
However, there is only 1 list of overall comma positions in the CST
node. Hence we implement a way to split the commas so that we can
separate `generics` and `genericPacks`. Note that this means the last
element of `generics` in the punctuated list may still have a comma.

---------

Co-authored-by: ariel <aweiss@hey.com>