CST: Improve serialization of functions, handle AstTypeFunction (#275)
This PR introduces support for serializing an AstTypeFunction. The most
interesting part of this is the parameters, where the name (and hence
colon) is optional. For this, we introduce a type
`AstTypeFunctionParameter` to contain this information. We also
serialize the vararg annotation independently.

As part of this, we also extend the serialization on normal functions to
include attributes, varargs, and the return type.

For now we serialize a placeholder for the `:` token (in functions and
for AstLocal too), which will get fixed after we sync to Luau 0.673.