Remove the misinformation in @batteries/pp's FIXME comments. (#282)
Someone tried to cite this incorrect comment as an indication that Luau
would eventually do something unsound, which will not happen. When I
wrote the comment, I was thinking about read-only indexers specifically,
but we don't have syntax or support for read-only indexers yet, and the
relationship implied by the comment doesn't even really hold for a
read-only `unknown` indexer, but is at least _safe_ (hence why the cast
we have today is alright because we don't mutate the table, we're just
reading it).