Simplify CST node existence checking during serialization (#289)
After a successful parse, if `options.storeCstData` is enabled, then a
CST node is guaranteed to exist containing the necessary position
information for serialization.

This PR simplifies our serialization logic to remove checking for a CST
node. `lookupCstNode` now returns a `Luau::NotNull<CstNode>` instead.

The result of serialization now matches the types we've defined in
`ast_types.luau`. Note that we do not serialize an AST anymore: instead,
a user can abstract from CST -> AST if their use case requires an AST
instead.

There is one case that this does not hold true: an AstTypeReference for
the reference `nil`. A corresponding CST node does not exist in this
case. So, for this, we gated the cst node lookup behind a check to see
if `prefix` exists or `hasParameterList` is true - both cases where we
need to access the CST node data.