Implement `Runtime::runOnce()` and use it in the internal scheduler loop (#286)
This PR implements a C-sided API `Runtime::runOnce()` to run a singular
iteration of the scheduler loop. This returns an `std::variant` between
`StepErr` and `StepSuccess`, returning the `lua_State* L` which errored
if the variant is a `StepErr`. This is then utilized in
`Runtime::runToCompletion`, which simplifies some logic regarding
exiting the runtime loop.

Fixes #272

---------

Co-authored-by: ariel <aweiss@hey.com>