Fix functions returned from `luau.load` not being able to call `require` (#301)
`luau.load` currently doesn't pass a valid chunkname to `luau_load` that
can be used to handle calls to `require`. This PR fixes that by
retrieving the source path of the caller using `lua_getinfo`. If no
second argument was passed and if the source path of the caller couldn't
be retrieved, `luau.load` will throw an error.

This PR also changes `luau.load` so that the second argument is
considered a path, relative to the current working directory. This path
is converted to a proper chunk name by inserting '@' at the start.