Implement password submodule for `@lute/crypto`. (#319)
As part of #161, implements a password submodule of `@lute/crypto` with
`hash` and `verify` functions that allow you to appropriately hash and
verify passwords securely. The underlying implementation leverages
libsodium because boringssl does not feature any appropriate key
derivation functions. We will continue to prefer boringssl cryptography
to libsodium for digest and other additional functionality for the
library.