Hash tune files to avoid unnecessary re-fetch when configuring (#326)
When we call `configure` in both versions of luthier, we now check if
our `.tune` files are out-of-date and re-fetch only if necessary. The
`fetch` subcommand always performs a fetch.

In the process, I also discovered why `luthier.luau` and `luthier.py`
were spitting out different hashes. It wasn't a discrepancy in the order
of files being processed—it was the digest size. I've made these equal
now, so both tools should be compatible again.

Resolves #226.