Remove luthier.py since it's been superseded entirely by the Luau implementation. (#337)
We no longer need to maintain `luthier.py`, and it poses a frustrating
maintenance risk right now to attempt to apply both together. It may be
useful in the future to introduce an optional path in the CI to fetch
everything from a tarball somewhere and then use cmake/ninja directly
instead of running luthier, but we can deal with that when needed.